backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.3
pyjwt==2.8.1
python-multipart==0.0.6
supabase==2.3.5
